+++ versions.tf
@@ -3,7 +3,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 3.34, < 4.0.0"
+      version = ">= 3.47, < 4.0.0"
     }
   }
 }